requirements.txt
# =============================================
# Python依赖清单
# =============================================
# 项目: 智能定时资讯推送系统
# Python版本: 3.11+
# =============================================

# GLM API客户端
zhipuai>=2.0.0

# HTTP请求
requests>=2.31.0

# YAML配置文件解析
PyYAML>=6.0.1

# 环境变量管理（可选，本地开发用）
python-dotenv>=1.0.0

# RSS订阅解析（新闻收集）
feedparser>=6.0.10

# 日期时间解析增强
python-dateutil>=2.8.2

# 类型提示（可选，代码质量工具）
typing-extensions>=4.8.0

# =============================================
# 可选依赖（未来扩展）
# =============================================

# Markdown转HTML（邮件推送优化）
# markdown>=3.5

# 任务调度（本地定时任务）
# schedule>=1.2.0

# 数据处理（高级去重）
# pandas>=2.1.0

# 日志增强
# loguru>=0.7.0

# =============================================
# 开发依赖（本地开发时安装）
# =============================================
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.10.0
# flake8>=6.1.0
# mypy>=1.6.0
